lua/aw_watcher/init.lua: send heartbeat on BufLeave
When switching buffers, watcher should sent 2 events to aw server,
one for leaving buffer and the other for entering. Right now event
is only sent on buffer entry which results in data loss. Consider
the case:

1. modify file A (event is sent for file A)
2. ponder for a minute without typing anything
3. switch buffer to B (only event for file B is sent now)

Now the whole minute of working on A is gone.

By sending event for A then B, this will make sure that this one
minute is recorded and not lost.

Signed-off-by: Michał Łyszczek <michal.lyszczek@bofc.pl>